+++ src/Cocoar.SignalARRR.Server/IClientContextDispatcher.cs
@@ -6,8 +6,6 @@ using Cocoar.SignalARRR.Common;
 namespace Cocoar.SignalARRR.Server {
     internal interface IClientContextDispatcher {
 
-        //Task ProxyClientAsync(string clientId, ServerRequestMessage serverRequestMessage, HttpContext httpContext);
-
         Task<TResult> InvokeClientAsync<TResult>(string clientId, ServerRequestMessage serverRequestMessage,
             CancellationToken cancellationToken);